refactor: extract shared calculate_refund_credits utility
Addresses Copilot code review feedback to extract the duplicated
_calculate_refund_credits function from both equipment/removal.py
and fighter/removal.py into a shared refund.py module.
