+++ gyrinx/core/handlers/equipment/removal.py
@@ -18,6 +18,7 @@ from gyrinx.content.models import (
     ContentWeaponProfile,
     VirtualWeaponProfile,
 )
+from gyrinx.core.handlers.refund import calculate_refund_credits
 from gyrinx.core.models.action import ListAction, ListActionType
 from gyrinx.core.models.list import (
     List,
@@ -49,30 +50,6 @@ class EquipmentComponentRemovalResult:
     refund_applied: bool
     description: str
     list_action: Optional[ListAction]
-
-
-def _calculate_refund_credits(
-    *,
-    lst: List,
-    cost: int,
-    request_refund: bool,
-) -> tuple[int, bool]:
-    """
-    Calculate credits delta based on refund request and campaign mode.
-
-    Refunds are ONLY allowed in campaign mode.
-
-    Args:
-        lst: The list
-        cost: The item cost being removed
-        request_refund: Whether user requested refund
-
-    Returns:
-        Tuple of (credits_delta, refund_applied)
-    """
-    refund_applied = request_refund and lst.is_campaign_mode
-    credits_delta = cost if refund_applied else 0
-    return credits_delta, refund_applied
 
 
 @transaction.atomic
@@ -120,7 +97,7 @@ def handle_equipment_removal(
     stash_delta = -equipment_cost if is_stash else 0
 
     # Validate and calculate refund
-    credits_delta, refund_applied = _calculate_refund_credits(
+    credits_delta, refund_applied = calculate_refund_credits(
         lst=lst,
         cost=equipment_cost,
         request_refund=request_refund,
@@ -226,7 +203,7 @@ def handle_equipment_component_removal(
     stash_delta = -component_cost if is_stash else 0
 
     # Validate and calculate refund
-    credits_delta, refund_applied = _calculate_refund_credits(
+    credits_delta, refund_applied = calculate_refund_credits(
         lst=lst,
         cost=component_cost,
         request_refund=request_refund,
